# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Builders MVP API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "post-controller"
                ],
                "operationId": "getPosts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    },
                    {
                        "name": "sortType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "LATEST",
                            "enum": [
                                "LATEST",
                                "MOST_VIEWED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePostListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "게시글 목록 조회 (페이지네이션, 정렬: 최신순/조회순)",
                "summary": "게시글 목록 조회 (페이지네이션, 정렬: 최신순/조회순)"
            },
            "post": {
                "tags": [
                    "post-controller"
                ],
                "operationId": "createPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTherapyPostRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapyPostDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "게시글 작성 (COMMUNITY 또는 RESOURCE 타입)",
                "summary": "게시글 작성 (COMMUNITY 또는 RESOURCE 타입)"
            }
        },
        "/api/v1/posts/{postId}/scrap": {
            "get": {
                "tags": [
                    "scrap-controller"
                ],
                "operationId": "getScrapStatus",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseScrapStatusResponse"
                                }
                            }
                        }
                    }
                },
                "description": "게시글 스크랩 여부 확인",
                "summary": "게시글 스크랩 여부 확인"
            },
            "post": {
                "tags": [
                    "scrap-controller"
                ],
                "operationId": "addScrap",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseScrapStatusResponse"
                                }
                            }
                        }
                    }
                },
                "description": "게시글 스크랩 추가",
                "summary": "게시글 스크랩 추가"
            },
            "delete": {
                "tags": [
                    "scrap-controller"
                ],
                "operationId": "removeScrap",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseScrapStatusResponse"
                                }
                            }
                        }
                    }
                },
                "description": "게시글 스크랩 삭제",
                "summary": "게시글 스크랩 삭제"
            }
        },
...
        "/api/v1/posts/{postId}": {
            "get": {
                "tags": [
                    "post-controller"
                ],
                "operationId": "getPostDetail",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapyPostDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "게시글 상세 조회 (조회수 증가, 리액션/스크랩 상태 포함)",
                "summary": "게시글 상세 조회 (조회수 증가, 리액션/스크랩 상태 포함)"
            },
            "delete": {
                "tags": [
                    "post-controller"
                ],
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "게시글 삭제 (소프트 삭제, 작성자 본인만 가능)",
                "summary": "게시글 삭제 (소프트 삭제, 작성자 본인만 가능)"
            },
            "patch": {
                "tags": [
                    "post-controller"
                ],
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTherapyPostRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapyPostDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "게시글 수정 (제목/내용/치료영역/연령대, 작성자 본인만 가능)",
                "summary": "게시글 수정 (제목/내용/치료영역/연령대, 작성자 본인만 가능)"
            }
        },
...
        "/api/v1/me/scraps": {
            "get": {
                "tags": [
                    "scrap-controller"
                ],
                "operationId": "getMyScraps",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseScrapListResponse"
                                }
                            }
                        }
                    }
                },
                "description": "내 스크랩 목록 조회 (페이지네이션)",
                "summary": "내 스크랩 목록 조회 (페이지네이션)"
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateTherapyPostRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "postType": {
                        "type": "string",
                        "enum": [
                            "COMMUNITY",
                            "RESOURCE"
                        ]
                    },
                    "therapyArea": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "OCCUPATIONAL",
                            "SPEECH",
                            "COGNITIVE",
                            "PLAY"
                        ]
                    },
                    "ageGroup": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "AGE_0_2",
                            "AGE_3_5",
                            "AGE_6_12",
                            "AGE_13_18",
                            "AGE_19_64",
                            "AGE_65_PLUS"
                        ]
                    }
                },
                "required": [
                    "ageGroup",
                    "content",
                    "postType",
                    "therapyArea",
                    "title"
                ]
            },
            "ApiResponseTherapyPostDetailResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TherapyPostDetailResponse"
                    }
                }
            },
            "PostAttachmentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "originalFilename": {
                        "type": "string"
                    },
                    "contentType": {
                        "type": "string"
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "downloadUrl": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "TherapyPostDetailResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "postType": {
                        "type": "string",
                        "enum": [
                            "COMMUNITY",
                            "RESOURCE"
                        ]
                    },
                    "authorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "authorNickname": {
                        "type": "string"
                    },
                    "therapyArea": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "OCCUPATIONAL",
                            "SPEECH",
                            "COGNITIVE",
                            "PLAY"
                        ]
                    },
                    "ageGroup": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "AGE_0_2",
                            "AGE_3_5",
                            "AGE_6_12",
                            "AGE_13_18",
                            "AGE_19_64",
                            "AGE_65_PLUS"
                        ]
                    },
                    "viewCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "canEdit": {
                        "type": "boolean"
                    },
                    "canDelete": {
                        "type": "boolean"
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostAttachmentResponse"
                        }
                    }
                }
            },
            "ApiResponseScrapStatusResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ScrapStatusResponse"
                    }
                }
            },
            "ScrapStatusResponse": {
                "type": "object",
                "properties": {
                    "postId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "scrapped": {
                        "type": "boolean"
                    }
                }
            },
...
            "UpdateTherapyPostRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "therapyArea": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "OCCUPATIONAL",
                            "SPEECH",
                            "COGNITIVE",
                            "PLAY"
                        ]
                    },
                    "ageGroup": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "AGE_0_2",
                            "AGE_3_5",
                            "AGE_6_12",
                            "AGE_13_18",
                            "AGE_19_64",
                            "AGE_65_PLUS"
                        ]
                    }
                },
                "required": [
                    "ageGroup",
                    "content",
                    "therapyArea",
                    "title"
                ]
            },
...
            "ApiResponsePostListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PostListResponse"
                    }
                }
            },
            "PostListResponse": {
                "type": "object",
                "properties": {
                    "posts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TherapyPostSummaryResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "hasNext": {
                        "type": "boolean"
                    }
                }
            },
            "TherapyPostSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "postType": {
                        "type": "string",
                        "enum": [
                            "COMMUNITY",
                            "RESOURCE"
                        ]
                    },
                    "title": {
                        "type": "string"
                    },
                    "contentPreview": {
                        "type": "string"
                    },
                    "authorNickname": {
                        "type": "string"
                    },
                    "therapyArea": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "OCCUPATIONAL",
                            "SPEECH",
                            "COGNITIVE",
                            "PLAY"
                        ]
                    },
                    "ageGroup": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "AGE_0_2",
                            "AGE_3_5",
                            "AGE_6_12",
                            "AGE_13_18",
                            "AGE_19_64",
                            "AGE_65_PLUS"
                        ]
                    },
                    "viewCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseScrapListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ScrapListResponse"
                    }
                }
            },
            "ScrapListResponse": {
                "type": "object",
                "properties": {
                    "scraps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ScrappedPostResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "hasNext": {
                        "type": "boolean"
                    }
                }
            },
            "ScrappedPostResponse": {
                "type": "object",
                "properties": {
                    "postId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "contentPreview": {
                        "type": "string"
                    },
                    "authorNickname": {
                        "type": "string"
                    },
                    "therapyArea": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "OCCUPATIONAL",
                            "SPEECH",
                            "COGNITIVE",
                            "PLAY"
                        ]
                    },
                    "ageGroup": {
                        "type": "string",
                        "enum": [
                            "UNSPECIFIED",
                            "AGE_0_2",
                            "AGE_3_5",
                            "AGE_6_12",
                            "AGE_13_18",
                            "AGE_19_64",
                            "AGE_65_PLUS"
                        ]
                    },
                    "viewCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "postCreatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "scrappedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}